chain: scale game money 1:1 to PARK ($1 = 1 PARK), widen amount to uint128
Bumps kGameMoneyToWei from 10^15 to 10^17 so explorers show readable
PARK quantities that track in-game prices 1:1: $10 entry → 10 PARK,
$5 ride → 5 PARK, $50 starting cash → 50 PARK.

The new scale exceeds uint64 for any plausible game amount (50 PARK is
already 5×10^19 wei vs uint64 max 1.84×10^19), so the Outbox amount
field widens from uint64_t to unsigned __int128. Adds AppendUnsigned128
for decimal serialization (snprintf has no format specifier for it).
The wire format is unchanged — amounts already serialize as JSON
strings and the sidecar parser already uses bigint, so the path round-
trips end-to-end.

outbox_roundtrip now pushes a 50 PARK GUEST_ENTRY (5×10^19 wei) as a
regression check that values overflowing the old uint64 path still
serialize and parse correctly. 8/8 lines OK.

+++ src/openrct2/chain/Outbox.h
@@ -111,13 +111,15 @@ namespace OpenRCT2::Chain
         // -------- Producer API (single-thread; game tick) --------
         // `cashWei` and `amountWei` are integer wei values; the caller converts in-game
         // money to wei. Sidecar parses these as bigint per its uint256 ABI contract.
+        // `unsigned __int128` covers the full range needed at kGameMoneyToWei = 10^17 —
+        // even the wealthiest guest's CashInPocket fits comfortably below 2^128.
 
-        void PushGuestEntry(int32_t guestId, uint32_t hdIndex, uint64_t cashWei);
+        void PushGuestEntry(int32_t guestId, uint32_t hdIndex, unsigned __int128 cashWei);
         void PushGuestSpend(
             int32_t guestId,
             uint32_t hdIndex,
             uint32_t venueId,
-            uint64_t amountWei,
+            unsigned __int128 amountWei,
             SpendCategory category,
             uint64_t gameTick);
         void PushGuestExit(int32_t guestId, uint32_t hdIndex);
